refactory unit tests without create real data on database

+++ src/main/kotlin/io/github/controleagenda/services/impl/SubSegmentServiceImpl.kt
@@ -1,6 +1,7 @@
 package io.github.controleagenda.services.impl
 
 import io.github.controleagenda.exception.BackendException
+import io.github.controleagenda.exception.NotFoundException
 import io.github.controleagenda.model.Segment
 import io.github.controleagenda.model.SegmentToReturn
 import io.github.controleagenda.model.SubSegment
@@ -24,38 +25,46 @@ class SubSegmentServiceImpl : SubSegmentService {
     val util = Util()
 
     override fun createSubSegment(idSegment: Long, subSegment: SubSegment): SegmentToReturn {
-        val segment: Segment = segmentRepository.getById(idSegment)
-        if (subSegmentRepository.findSubSegmentToSegmentID(segment.id!!).count() < 20) {
-            subSegmentRepository.save(
-                SubSegment(
-                    util.idSequenceSubSegment(subSegmentRepository), subSegment.subSegmentName, subSegment.message,
-                    Segment(segment.id, segment.segmentName)
+
+        if (segmentRepository.findById(idSegment).isPresent) {
+            val segment: Segment = segmentRepository.getById(idSegment)
+            if (subSegmentRepository.findSubSegmentToSegmentID(segment.id!!).count() < 20) {
+                subSegmentRepository.save(
+                    SubSegment(
+                        util.idSequenceSubSegment(subSegmentRepository), subSegment.subSegmentName, subSegment.message,
+                        Segment(segment.id, segment.segmentName)
+                    )
                 )
-            )
-            val subSegments = subSegmentRepository.findAll()
-            subSegments.add(subSegment)
-            val subSegmentToSegment =
-                subSegments.filter {
-                    it!!.segment.id == (segment.id)
-                } as MutableList<SubSegment>
-            return SegmentToReturn(
-                Segment(segment.id, segment.segmentName),
-                subSegmentToSegment.toMutableList()
-            )
-        } else throw BackendException("Excedeu a quantidade de 20 Tarefas por Segmento")
+                val subSegments = subSegmentRepository.findAll()
+
+                val subSegmentToSegment =
+                    subSegments.filter {
+                        it!!.segment.id == (segment.id)
+                    } as MutableList<SubSegment>
+                return SegmentToReturn(
+                    Segment(segment.id, segment.segmentName),
+                    subSegmentToSegment.toMutableList()
+                )
+            } else throw BackendException("Excedeu a quantidade de 20 Tarefas por Segmento")
+        } else throw NotFoundException("Segmento com o id ${idSegment} não existe no banco de dados")
     }
 
     override fun updateSubSegment(subSegment: SubSegment): SubSegment {
-        val subSegmentBase = subSegmentRepository.findSubSegmentById(subSegment.id!!)
-        return subSegmentRepository.save(
-            SubSegment(
-                subSegment.id, subSegment.subSegmentName, subSegment.message,
-                Segment(subSegmentBase!!.segment.id, subSegmentBase.subSegmentName)
+        if (subSegmentRepository.findById(subSegment.id!!).isPresent) {
+            val subSegmentBase = subSegmentRepository.findSubSegmentById(subSegment.id!!)
+            return subSegmentRepository.save(
+                SubSegment(
+                    subSegment.id, subSegment.subSegmentName, subSegment.message,
+                    Segment(subSegmentBase!!.segment.id, subSegmentBase.subSegmentName)
+                )
             )
-        )
+        } else throw NotFoundException("SubSegmento com o id ${subSegment.id} não existe no banco de dados")
     }
 
     override fun deleteSubSegment(idSubSegment: Long) {
-        subSegmentRepository.deleteById(idSubSegment)
+        if (subSegmentRepository.findById(idSubSegment).isPresent) {
+            subSegmentRepository.deleteById(idSubSegment)
+        } else throw NotFoundException("SubSegmento com o id ${idSubSegment} não existe no banco de dados")
+
     }
 }
